requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pypdf>=4.0.0
pdf2image==1.16.3
pydantic==2.5.0
numpy<2.0.0
opencv-python==4.8.1.78
# torch và torchvision được cài riêng trong Dockerfile với CUDA support
easyocr>=1.7.0